refactor(InputState): InputState 위치 변경
